Made some changes for management meeting
I added the sprite and normal map to the player, which is now a little
bigger. I also added a little dynamic lights demo to the start menu and
added the scenes to the build settings so they can be played in
succession (although I don't know if that will stick, we may just need
to manually set those until we start making builds of the game.